requirements.txt
Django
django-crispy-forms
django-evolution
django-extensions
django-firebase-auth
django-jazzmin
djangorestframework
djangorestframework-api-key
Flask
mysql-connector
mysql-connector-python
mysqlclient
psycopg2-binary>=2.8,<3.0
neo4j
py2neo
django_neomodel
django-rest-swagger
celery
reddis